requirements.txt
torch>=2.0.0
escnn>=0.1.9
opencv-python>=4.7.0
numpy>=1.24.0
matplotlib>=3.6.0
pyyaml>=6.0

scipy>=1.10.0

# Tactile sensor dependency (install from source)
git+https://github.com/joehjhuang/gs_sdk.git
